Nuget package, build, and facade
Added nuget packaging of trout library
updated build to package the trout nuget package
Facade for easily creating getting an instance of MailMessageQueue and
MailMessageDequeuer.
Moved IoC to console, wasn't necessary or used in base library. Decision
on how to instantiate objects is left up to consumer of the code