Removed hard dependency on EF from core library:
Removed the EF depenency from the trout project so that trout can be used
with any persistency strategy (whether EF, NHibernate, or any other
persitence mechanism). Hard dependency on EF was really bad, especially
when attempting to use trout in a project that uses a different version of
EF.

Placed the EF implementation of IRepository in the console project. I
could argue that EF shouldnt be in this project either to support the
usability of the console project. It is okay for now since the core
assembly is what is needed the most.